pyproject.toml
[project]
name = "assembly_uploader"
readme = "README.md"
authors = [
  {name = "MGnify team", email = "metagenomics-help@ebi.ac.uk"},
]
license = {text = "Apache Software License 2.0"}
keywords = ["bioinformatics", "tool", "metagenomics"]
dynamic = ["version"]
description = "Python scripts to upload primary metagenome and metatranscriptome assemblies to ENA on a per-study basis. This script generates xmls to register a new study and create manifests necessary for submission with webin-cli."
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]

dependencies = [
    "requests==2.32.3"
]

[build-system]
requires = ["setuptools>=61.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = {attr = "assembly_uploader.__version__"}

[tool.setuptools]
packages = ["assembly_uploader"]

[project.scripts]
study_xmls = "assembly_uploader.study_xmls:main"
submit_study = "assembly_uploader.submit_study:main"
assembly_manifest = "assembly_uploader.assembly_manifest:main"
release_study = "assembly_uploader.release_study:main"

[project.optional-dependencies]
dev = [
    "pre-commit==3.6.0",
    "black==24.4.2",
]
test = [
    "pytest==8.2.2",
    "pytest-md==0.2.0",
    "pytest-workflow==2.1.0",
    "pytest-responses==0.5.1"
]

[tool.isort]
profile = "black"
